// client/src/widgets/gost-form/newGostModel.ts
export type GostToSave = {
	designation: string;
	fullName: string;
	codeOks: string;
	activityField: string;
	acceptanceYear: string;
	commissionYear: string;
	author: string;
	acceptedFirstTimeOrReplaced: string;
	content: string;
	keyWords: string;
	applicationArea: string;
	adoptionLevel: number;
	documentText: string;
	changes: string;
	amendments: string;
	status: number;
	harmonization: number;
	isPrimary: boolean;
	references: string[];
};